requirements.txt
alibabacloud-credentials>=0.2.0
alibabacloud-dingtalk>=1.2.18
alibabacloud-endpoint-util>=0.0.3
alibabacloud-gateway-spi>=0.0.1
alibabacloud-openapi-util>=0.1.6
alibabacloud-tea>=0.2.8
alibabacloud-tea-openapi>=0.3.1
alibabacloud-tea-util>=0.3.5
APScheduler>=3.8.1
beautifulsoup4>=4.10.0
httpx>=0.21.1
nb-cli>=0.5.1
nonebot-adapter-cqhttp>=2.0.0a16
nonebot-bison>=0.4.2
nonebot-plugin-status>=0.2.2
nonebot-plugin-apscheduler>=0.1.2
nonebot2>=2.0.0a16
mariadb>=1.0.9
